+++ common/protocol.h
@@ -163,14 +163,14 @@ typedef struct netMsg_s {
 	qBool		allowOverflow;	// if false, do a Com_Error
 	qBool		overFlowed;		// set to true if the buffer size failed
 	byte		*data;
-	int			maxSize;
-	int			curSize;
-	int			readCount;
-	int			bufferSize;
+	size_t		maxSize;
+	size_t		curSize;
+	size_t		readCount;
+	size_t		bufferSize;
 } netMsg_t;
 
 // supporting functions
-void	MSG_Init (netMsg_t *dest, byte *data, int length);
+void	MSG_Init (netMsg_t *dest, byte *data, size_t length);
 void	MSG_Clear (netMsg_t *dest);
 
 // writing
@@ -187,7 +187,7 @@ void	MSG_WriteDir (netMsg_t *dest, vec3_t vector);
 void	MSG_WriteFloat (netMsg_t *dest, float f);
 void	MSG_WriteInt3 (netMsg_t *dest, int c);
 void	MSG_WriteLong (netMsg_t *dest, int c);
-void	MSG_WriteRaw (netMsg_t *dest, void *data, int length);
+void	MSG_WriteRaw (netMsg_t *dest, void *data, size_t length);
 void	MSG_WriteShort (netMsg_t *dest, int c);
 void	MSG_WriteString (netMsg_t *dest, char *s);
 void	MSG_WriteStringCat (netMsg_t *dest, char *data);
@@ -201,7 +201,7 @@ void	MSG_WriteStringCat (netMsg_t *dest, char *data);
 void	MSG_BeginReading (netMsg_t *src);
 int		MSG_ReadByte (netMsg_t *src);
 int		MSG_ReadChar (netMsg_t *src);
-void	MSG_ReadData (netMsg_t *src, void *buffer, int size);
+void	MSG_ReadData (netMsg_t *src, void *buffer, size_t size);
 void	MSG_ReadDeltaUsercmd (netMsg_t *src, struct userCmd_s *from, struct userCmd_s *cmd);
 void	MSG_ReadDir (netMsg_t *src, vec3_t vector);
 float	MSG_ReadFloat (netMsg_t *src);
@@ -318,7 +318,7 @@ void		NET_Shutdown (void);
 netConfig_t NET_Config (netConfig_t openFlags);
 
 qBool		NET_GetPacket (netSrc_t sock, netAdr_t *fromAddr, netMsg_t *message);
-int			NET_SendPacket (netSrc_t sock, int length, void *data, netAdr_t *to);
+int			NET_SendPacket (netSrc_t sock, size_t length, void *data, netAdr_t *to);
 
 char		*NET_AdrToString (netAdr_t *a);
 qBool		NET_StringToAdr (char *s, netAdr_t *a);
@@ -354,14 +354,14 @@ typedef struct netChan_s {
 	byte		messageBuff[MAX_CL_USABLEMSG];	// leave space for header
 
 	// Message is copied to this buffer when it is first transfered
-	int			reliableLength;
+	size_t		reliableLength;
 	byte		reliableBuff[MAX_CL_USABLEMSG];	// unacked reliable message
 } netChan_t;
 
 void		Netchan_Init (void);
 void		Netchan_Setup (netSrc_t sock, netChan_t *chan, netAdr_t *adr, int protocol, int qPort, uint32 msgLen);
 
-int			Netchan_Transmit (netChan_t *chan, int length, byte *data);
-void		Netchan_OutOfBand (netSrc_t netSocket, netAdr_t *adr, int length, byte *data);
+int			Netchan_Transmit (netChan_t *chan, size_t length, byte *data);
+void		Netchan_OutOfBand (netSrc_t netSocket, netAdr_t *adr, size_t length, byte *data);
 void		Netchan_OutOfBandPrint (netSrc_t netSocket, netAdr_t *adr, char *format, ...);
 qBool		Netchan_Process (netChan_t *chan, netMsg_t *msg);
